# threadpilot/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ThreadPilot AI API",
    "description": "MySQL-grounded intent workflow, incremental synchronization and read-only SQL Agent.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local development; replace with your team server when deployed"
    }
  ],
  "paths": {
...
    "/api/v1/workflow/chat": {
      "post": {
        "tags": [
          "Workflow"
        ],
        "summary": "Workflow Chat",
        "description": "场景 1.1–3.3：JSON 工作流入口，传回 session_id 延续上下文。",
        "operationId": "workflow_chat_api_v1_workflow_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chat": {
      "post": {
        "tags": [
          "Workflow"
        ],
        "summary": "Workflow Chat",
        "description": "场景 1.1–3.3：JSON 工作流入口，传回 session_id 延续上下文。",
        "operationId": "workflow_chat_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflow/chat/stream": {
      "post": {
        "tags": [
          "Workflow"
        ],
        "summary": "Workflow Stream",
        "description": "场景 1.1–3.3：完成规则验证后发送 SSE，避免未验证文本抢先展示。",
        "operationId": "workflow_stream_api_v1_workflow_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Validated workflow SSE: start, delta, done",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evidence/{source}/{row}": {
      "get": {
        "tags": [
          "Evidence"
        ],
        "summary": "Evidence Row",
        "description": "场景 1.3/2.1/2.2：白名单访问原始记录，展示引用对应数据。",
        "operationId": "evidence_row_api_v1_evidence__source___row__get",
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source"
            }
          },
          {
            "name": "row",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Row"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evidence"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflow/notifications/{session_id}": {
      "get": {
        "tags": [
          "Workflow"
        ],
        "summary": "Notifications",
        "description": "场景 3.3：读取已评估的站内通知；会话 ID 是本地单用户访问凭证。",
        "operationId": "notifications_api_v1_workflow_notifications__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Notifications Api V1 Workflow Notifications  Session Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflow/replies": {
      "post": {
        "tags": [
          "Integrations"
        ],
        "summary": "Receive Reply",
        "description": "场景 3.3：已认证上游回复事件入库；聊天消息不能伪造已收到回复。",
        "operationId": "receive_reply_api_v1_workflow_replies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Receive Reply Api V1 Workflow Replies Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/evidence/{dataset}/{record_id}": {
      "get": {
        "tags": [
          "Data"
        ],
        "summary": "Evidence",
        "description": "场景 1.3/2.2：展示查询时的当前记录、来源、时间和版本。",
        "operationId": "evidence_api_evidence__dataset___record_id__get",
        "parameters": [
          {
            "name": "dataset",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "orders",
                "production_log",
                "workshops"
              ],
              "type": "string",
              "title": "Dataset"
            }
          },
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Evidence Api Evidence  Dataset   Record Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/import": {
      "post": {
        "tags": [
          "Synchronization"
        ],
        "summary": "Import File",
        "description": "场景 1.2：上传后原子导入，不覆盖原始文件，不接受服务器任意路径。",
        "operationId": "import_file_api_sync_import_post",
        "parameters": [
          {
            "name": "dataset",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "orders",
                    "production_log",
                    "workshops"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dataset"
            }
          },
          {
            "name": "x-confirm-write",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Confirm-Write"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_file_api_sync_import_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/logs": {
      "get": {
        "tags": [
          "Synchronization"
        ],
        "summary": "Logs",
        "description": "场景 1.2/2.2：管理员读取最近一百条同步审计。",
        "operationId": "logs_api_sync_logs_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SyncRead"
                  },
                  "title": "Response Logs Api Sync Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_file_api_sync_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_file_api_sync_import_post"
      },
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 1500,
            "minLength": 1,
            "title": "Message"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-f0-9-]{36}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "confirmation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Id"
          },
          "selected_order_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^ORD-\\d{3}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Order Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "intent": {
            "$ref": "#/components/schemas/Intent"
          },
          "slots": {
            "$ref": "#/components/schemas/Slots"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "needs_clarification": {
            "type": "boolean",
            "title": "Needs Clarification",
            "default": false
          },
          "clarification_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clarification Question"
          },
          "confirmation_required": {
            "type": "boolean",
            "title": "Confirmation Required",
            "default": false
          },
          "confirmation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Id"
          },
          "state": {
            "$ref": "#/components/schemas/DialogState"
          },
          "tool_calls": {
            "items": {
              "$ref": "#/components/schemas/ToolCall"
            },
            "type": "array",
            "title": "Tool Calls"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "order_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Order Ids"
          },
          "selected_order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Order Id"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "gpt-4.1"
          },
          "business_date": {
            "type": "string",
            "title": "Business Date"
          },
          "usage": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "answer",
          "intent",
          "slots",
          "confidence",
          "state",
          "business_date"
        ],
        "title": "ChatResponse"
      },
      "DialogState": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "active_order": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Order"
          },
          "active_customer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Customer"
          },
          "pending_clarification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PendingClarification"
              },
              {
                "type": "null"
              }
            ]
          },
          "pending_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PendingAction"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_intent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Intent"
              },
              {
                "type": "null"
              }
            ]
          },
          "slots": {
            "$ref": "#/components/schemas/Slots"
          },
          "comparison_orders": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Comparison Orders"
          },
          "last_order_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Last Order Ids"
          },
          "draft": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft"
          },
          "history": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "History"
          },
          "revision": {
            "type": "integer",
            "title": "Revision",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DialogState"
      },
      "Evidence": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "row": {
            "type": "integer",
            "title": "Row"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "record": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Record"
          },
          "record_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Id"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "original_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "row",
          "url",
          "record"
        ],
        "title": "Evidence"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Intent": {
        "type": "string",
        "enum": [
          "order.lookup",
          "order.refresh",
          "order.risk",
          "order.compare",
          "order.commitment",
          "order.prioritize",
          "operations.normality",
          "operations.deviation",
          "execution.chase",
          "execution.note",
          "execution.reminder",
          "unknown"
        ],
        "title": "Intent"
      },
...
      "ReplyEvent": {
        "properties": {
          "event_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Event Id"
          },
          "order_id": {
            "type": "string",
            "pattern": "^ORD-\\d{3}$",
            "title": "Order Id"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "order_id",
          "received_at"
        ],
        "title": "ReplyEvent"
      },
...
      "SyncRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "file_hash": {
            "type": "string",
            "title": "File Hash"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "inserted": {
            "type": "integer",
            "title": "Inserted"
          },
          "updated": {
            "type": "integer",
            "title": "Updated"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "file_hash",
          "status",
          "inserted",
          "updated",
          "skipped",
          "message",
          "created_at"
        ],
        "title": "SyncRead"
      },
      "ToolCall": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "arguments": {
            "additionalProperties": true,
            "type": "object",
            "title": "Arguments"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "idempotent": {
            "type": "boolean",
            "title": "Idempotent",
            "default": true
          },
          "requires_confirmation": {
            "type": "boolean",
            "title": "Requires Confirmation",
            "default": false
          },
          "executed": {
            "type": "boolean",
            "title": "Executed",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ToolCall"
      },
...
    }
  }
}